Extract ToolsMenu and search-result navigation from TopBar
